Каркас шахматной доски[#93386] Рефакторинг кода. Добавление поведения коня.

+++ chess/src/main/java/ru/job4j/chess/firuges/BasicFigure.java
@@ -16,7 +16,7 @@ public abstract class BasicFigure implements Figure {
     @Override
     abstract public Cell[] way(Cell source, Cell dest) throws ImpossibleMoveException;
 
-    protected Cell findBy(double cellX, double cellY) {
+    protected Cell findBy(int cellX, int cellY) {
         Cell rst = Cell.A1;
         for (Cell cell : Cell.values()) {
             if ((cell.x ==cellX) && (cell.y == cellY)) {
